Engraving: move note input commands from Score into NoteInput
Absorb all of dom/noteentry.cpp plus the related note-entry methods from
edit.cpp and cmd.cpp (padToggle, addPitch/addNote/addMidiPitch and
friends, noteVal, noteValForPosition, putNote, repitchNote,
insertChordByInsertingTime, truncateChordRest, nextInputPos, cmdAddPitch,
cmdAddFret, resolveNoteInputParams) into a new static NoteInput class in
editing/, each new method taking a Transaction& parameter. Delete the now
empty noteentry.cpp.
