+++ src/engraving/dom/score.cpp
@@ -2824,312 +2824,6 @@ void Score::cmdConcertPitchChanged(bool flag)
     }
 }
 
-void Score::padToggle(Pad p, bool toggleForSelectionOnly)
-{
-    if (!noteEntryMode()) {
-        for (ChordRest* cr : getSelectedChordRests()) {
-            if (cr->isChord() && toChord(cr)->isTrillCueNote()) {
-                return;
-            }
-        }
-    }
-
-    const TDuration oldDuration = m_is.duration();
-    const bool oldRest = m_is.rest();
-    const AccidentalType oldAccidentalType = m_is.accidentalType();
-
-    switch (p) {
-    case Pad::NOTE00:
-        m_is.setDuration(DurationType::V_LONG);
-        break;
-    case Pad::NOTE0:
-        m_is.setDuration(DurationType::V_BREVE);
-        break;
-    case Pad::NOTE1:
-        m_is.setDuration(DurationType::V_WHOLE);
-        break;
-    case Pad::NOTE2:
-        m_is.setDuration(DurationType::V_HALF);
-        break;
-    case Pad::NOTE4:
-        m_is.setDuration(DurationType::V_QUARTER);
-        break;
-    case Pad::NOTE8:
-        m_is.setDuration(DurationType::V_EIGHTH);
-        break;
-    case Pad::NOTE16:
-        m_is.setDuration(DurationType::V_16TH);
-        break;
-    case Pad::NOTE32:
-        m_is.setDuration(DurationType::V_32ND);
-        break;
-    case Pad::NOTE64:
-        m_is.setDuration(DurationType::V_64TH);
-        break;
-    case Pad::NOTE128:
-        m_is.setDuration(DurationType::V_128TH);
-        break;
-    case Pad::NOTE256:
-        m_is.setDuration(DurationType::V_256TH);
-        break;
-    case Pad::NOTE512:
-        m_is.setDuration(DurationType::V_512TH);
-        break;
-    case Pad::NOTE1024:
-        m_is.setDuration(DurationType::V_1024TH);
-        break;
-    case Pad::REST:
-        if (noteEntryMode()) {
-            m_is.setRest(!m_is.rest());
-            m_is.setAccidentalType(AccidentalType::NONE);
-        } else if (selection().isNone()) {
-            m_is.setDuration(DurationType::V_QUARTER);
-            m_is.setRest(true);
-        } else {
-            for (ChordRest* cr : getSelectedChordRests()) {
-                if (!cr->isRest()) {
-                    setNoteRest(cr->segment(), cr->track(), NoteVal(), cr->durationTypeTicks(), DirectionV::AUTO, false,
-                                m_is.articulationIds());
-                }
-            }
-        }
-        break;
-    case Pad::DOT:
-        if ((m_is.duration().dots() == 1) || (m_is.duration() == DurationType::V_1024TH)) {
-            m_is.setDots(0);
-        } else {
-            m_is.setDots(1);
-        }
-        break;
-    case Pad::DOT2:
-        if ((m_is.duration().dots() == 2)
-            || (m_is.duration() == DurationType::V_512TH)
-            || (m_is.duration() == DurationType::V_1024TH)) {
-            m_is.setDots(0);
-        } else {
-            m_is.setDots(2);
-        }
-        break;
-    case Pad::DOT3:
-        if ((m_is.duration().dots() == 3)
-            || (m_is.duration() == DurationType::V_256TH)
-            || (m_is.duration() == DurationType::V_512TH)
-            || (m_is.duration() == DurationType::V_1024TH)) {
-            m_is.setDots(0);
-        } else {
-            m_is.setDots(3);
-        }
-        break;
-    case Pad::DOT4:
-        if ((m_is.duration().dots() == 4)
-            || (m_is.duration() == DurationType::V_128TH)
-            || (m_is.duration() == DurationType::V_256TH)
-            || (m_is.duration() == DurationType::V_512TH)
-            || (m_is.duration() == DurationType::V_1024TH)) {
-            m_is.setDots(0);
-        } else {
-            m_is.setDots(4);
-        }
-        break;
-    }
-    if (p >= Pad::NOTE00 && p <= Pad::NOTE1024) {
-        m_is.setDots(0);
-        //
-        // if in "note enter" mode, reset
-        // rest flag
-        //
-        if (noteEntryMode() && !toggleForSelectionOnly) {
-            if (usingNoteEntryMethod(NoteEntryMethod::BY_DURATION) || usingNoteEntryMethod(NoteEntryMethod::RHYTHM)) {
-                switch (oldDuration.dots()) {
-                case 1:
-                    padToggle(Pad::DOT);
-                    break;
-                case 2:
-                    padToggle(Pad::DOT2);
-                    break;
-                case 3:
-                    padToggle(Pad::DOT3);
-                    break;
-                case 4:
-                    padToggle(Pad::DOT4);
-                    break;
-                }
-
-                if (m_is.rest()) {
-                    // Enter a rest
-                    setNoteRest(m_is.segment(), m_is.track(), NoteVal(), m_is.duration().fraction());
-                    m_is.moveToNextInputPos();
-                } else {
-                    if (usingNoteEntryMethod(NoteEntryMethod::RHYTHM)) {
-                        const EngravingItem* selectedItem = score()->selection().element();
-                        if (selectedItem && selectedItem->isNote()) {
-                            m_is.setNotes({ toNote(selectedItem)->noteVal() });
-                        }
-                    }
-
-                    if (m_is.beyondScore()) {
-                        appendMeasures(1);
-                        m_is.moveToNextInputPos();
-                    }
-                    ChordRest* cr = m_is.cr();
-                    Chord* chord = nullptr;
-
-                    if (cr && cr->isChord() && cr->durationType() == m_is.duration()) {
-                        chord = toChord(cr);
-                    }
-
-                    for (const NoteVal& nval : m_is.notes()) {
-                        if (chord && chord->findNote(nval.pitch)) {
-                            m_is.moveToNextInputPos();
-                            continue;
-                        }
-
-                        NoteVal copy(nval);
-                        const Note* note = nullptr;
-
-                        if (chord) {
-                            note = addPitchToChord(copy, chord);
-                        } else {
-                            note = addPitch(copy, false /*addFlag*/);
-                        }
-
-                        if (note) {
-                            chord = note->chord();
-                        }
-                    }
-                }
-            } else {
-                m_is.setRest(false);
-            }
-        }
-    }
-
-    if (noteEntryMode()) {
-        if (!toggleForSelectionOnly || selection().isNone()) {
-            return;
-        }
-    }
-
-    std::vector<ChordRest*> crs;
-    std::vector<EngravingItem*> elementsToSelect;
-
-    if (selection().isSingle()) {
-        EngravingItem* e = selection().element();
-        ChordRest* cr = InputState::chordRest(e);
-
-        // do not allow to add a dot on a full measure rest
-        if (cr && cr->isRest()) {
-            Rest* r = toRest(cr);
-            if (r->isFullMeasureRest()) {
-                m_is.setDots(0);
-            }
-        }
-
-        // on measure rest, select the first actual rest
-        if (cr && cr->isMMRest()) {
-            Measure* m = cr->measure()->mmRestFirst();
-            if (m) {
-                cr = m->findChordRest(m->tick(), 0);
-            }
-        }
-
-        if (cr) {
-            crs.push_back(cr);
-        } else {
-            deselect(e);
-        }
-    } else if (selection().isNone() && p != Pad::REST) {
-        TDuration td = m_is.duration();
-        m_is.setDuration(td);
-        m_is.setAccidentalType(AccidentalType::NONE);
-    } else {
-        const auto elements = selection().uniqueElements();
-        for (EngravingItem* e : elements) {
-            if (selection().isList() || e->isRest() || e->isNote()) {
-                elementsToSelect.push_back(e);
-                deselect(e);
-            }
-        }
-        bool canAdjustLength = true;
-        for (EngravingItem* e : elements) {
-            ChordRest* cr = InputState::chordRest(e);
-            if (!cr) {
-                continue;
-            }
-            if (cr->isMeasureRepeat() || cr->isMMRest()) {
-                canAdjustLength = false;
-                break;
-            }
-            crs.push_back(cr);
-        }
-
-        if (canAdjustLength) {
-            // Change length from last to first chord/rest
-            std::sort(crs.begin(), crs.end(), [](const ChordRest* cr1, const ChordRest* cr2) {
-                if (cr2->track() == cr1->track()) {
-                    return cr2->isBefore(cr1);
-                }
-                return cr2->track() < cr1->track();
-            });
-            // Remove duplicates from the list
-            crs.erase(std::unique(crs.begin(), crs.end()), crs.end());
-        } else {
-            crs.clear();
-        }
-    }
-
-    const int dots = m_is.duration().dots();
-
-    if (toggleForSelectionOnly && dots > 0 && !crs.empty()) {
-        bool shouldRemoveDots = true;
-        for (const ChordRest* cr : crs) {
-            if (dots != cr->dots()) {
-                shouldRemoveDots = false;
-                break;
-            }
-        }
-
-        if (shouldRemoveDots) {
-            m_is.setDots(0);
-        }
-    }
-
-    for (ChordRest* cr : crs) {
-        if (cr->isChord() && (toChord(cr)->isGrace())) {
-            //
-            // handle appoggiatura and acciaccatura
-            //
-            undoChangeChordRestLen(cr, m_is.duration());
-        } else {
-            changeCRlen(cr, m_is.duration());
-        }
-    }
-
-    if (!elementsToSelect.empty()) {
-        std::vector<EngravingItem*> selectList;
-        for (EngravingItem* e : elementsToSelect) {
-            if (canReselectItem(e)) {
-                selectList.push_back(e);
-            }
-        }
-
-        select(selectList, SelectType::ADD, 0);
-        selection().updateSelectedElements();
-    }
-
-    if (toggleForSelectionOnly) {
-        m_is.setDuration(oldDuration);
-        m_is.setRest(oldRest);
-        m_is.setAccidentalType(oldAccidentalType);
-
-        if (noteEntryMode()) {
-            if (m_is.lastSegment() == m_is.segment()) {
-                m_is.moveToNextInputPos();
-            }
-        }
-    }
-}
-
 static void onFocusedItemChanged(EngravingItem* item)
 {
 #ifndef ENGRAVING_NO_ACCESSIBILITY
