+++ src/engraving/editing/edit.cpp
@@ -582,122 +582,6 @@ std::vector<Rest*> Score::setRests(const Fraction& _tick, track_idx_t track, con
     return rests;
 }
 
-//---------------------------------------------------------
-//   addNote from NoteVal
-//---------------------------------------------------------
-
-Note* Score::addNote(Chord* chord, const NoteVal& noteVal, bool forceAccidental, const std::set<SymId>& articulationIds,
-                     InputState* externalInputState)
-{
-    InputState& is = externalInputState ? (*externalInputState) : m_is;
-
-    Note* note = Factory::createNote(chord);
-    note->setParent(chord);
-    note->setTrack(chord->track());
-    note->setNval(noteVal);
-    undoAddElement(note);
-    if (forceAccidental) {
-        int tpc = style().styleB(Sid::concertPitch) ? noteVal.tpc1 : noteVal.tpc2;
-        AccidentalVal alter = tpc2alter(tpc);
-        AccidentalType at = Accidental::value2subtype(alter);
-        Accidental* a = Factory::createAccidental(note);
-        a->setAccidentalType(at);
-        a->setRole(AccidentalRole::USER);
-        a->setParent(note);
-        undoAddElement(a);
-    }
-
-    if (!articulationIds.empty()) {
-        chord->updateArticulations(articulationIds);
-    }
-
-    setPlayNote(true);
-    setPlayChord(true);
-
-    if (externalInputState) {
-        is.setTrack(note->track());
-        is.setLastSegment(is.segment());
-        is.setSegment(note->chord()->segment());
-    } else {
-        select(note, SelectType::SINGLE, 0);
-    }
-
-    if (!chord->staff()->isTabStaff(chord->tick())) {
-        NoteEntryMethod entryMethod = is.noteEntryMethod();
-        if (entryMethod != NoteEntryMethod::REPITCH && entryMethod != NoteEntryMethod::REALTIME_AUTO
-            && entryMethod != NoteEntryMethod::REALTIME_MANUAL) {
-            is.moveToNextInputPos();
-        }
-    }
-    return note;
-}
-
-Note* Score::addNoteToTiedChord(Chord* chord, const NoteVal& noteVal, bool forceAccidental, const std::set<SymId>& articulationIds)
-{
-    IF_ASSERT_FAILED(!chord->notes().empty()) {
-        return nullptr;
-    };
-    Note* referenceNote = chord->notes().at(0);
-
-    while (true) {
-        // don't add note if it is already part of tied notes previously
-        if (referenceNote->chord()->findNote(noteVal.pitch)) {
-            return nullptr;
-        }
-        if (!referenceNote->tieBack() || referenceNote->incomingPartialTie()) {
-            break;
-        }
-        referenceNote = referenceNote->tieBack()->startNote();
-    }
-
-    Tie* tie = nullptr;
-    Note* newNote = nullptr;
-
-    while (referenceNote->tieFor()) {
-        chord = referenceNote->chord();
-        newNote = addNote(chord, noteVal, forceAccidental, articulationIds);
-        if (!newNote) {
-            return nullptr;
-        }
-        if (tie) {
-            tie->setEndNote(newNote);
-            tie->setTick2(newNote->tick());
-            newNote->setTieBack(tie);
-            undoAddElement(tie);
-        }
-
-        tie = Factory::createTie(newNote);
-        tie->setStartNote(newNote);
-        tie->setTick(newNote->tick());
-        tie->setTrack(newNote->track());
-        newNote->setTieFor(tie);
-
-        if (!referenceNote->tieFor()->endNote()) {
-            break;
-        }
-
-        referenceNote = referenceNote->tieFor()->endNote();
-    }
-
-    chord = referenceNote->chord();
-    newNote = addNote(chord, noteVal, forceAccidental, articulationIds);
-
-    if (!newNote) {
-        return nullptr;
-    }
-
-    if (tie) {
-        tie->setEndNote(newNote);
-        tie->setTick2(newNote->tick());
-        newNote->setTieBack(tie);
-        undoAddElement(tie);
-    }
-
-    connectTies();
-
-    return newNote;
-}
-
 Slur* Score::addSlur(ChordRest* firstChordRest, ChordRest* secondChordRest, const Slur* slurTemplate)
 {
     if (!secondChordRest) {
@@ -1043,82 +927,6 @@ TextBase* Score::addText(TextStyleType type, EngravingItem* destinationElement)
     }
 
     return textBox;
-}
-
-//---------------------------------------------------------
-//  addTiedMidiPitch
-//---------------------------------------------------------
-
-Note* Score::addTiedMidiPitch(int pitch, bool addFlag, Chord* prevChord, bool allowTransposition)
-{
-    Note* n = addMidiPitch(pitch, addFlag, allowTransposition);
-    if (prevChord) {
-        Note* nn = prevChord->findNote(n->pitch());
-        if (nn) {
-            Tie* tie = Factory::createTie(this->dummy());
-            tie->setStartNote(nn);
-            tie->setEndNote(n);
-            tie->setTick(tie->startNote()->tick());
-            tie->setTick2(tie->endNote()->tick());
-            tie->setTrack(n->track());
-            n->setTieBack(tie);
-            nn->setTieFor(tie);
-            undoAddElement(tie);
-        }
-    }
-    return n;
-}
-
-NoteVal Score::noteVal(int pitch, staff_idx_t staffIdx, bool allowTransposition) const
-{
-    NoteVal nval(pitch);
-
-    const Staff* st = staff(staffIdx);
-    if (!st) {
-        return nval;
-    }
-
-    bool concertPitch = style().styleB(Sid::concertPitch);
-    Interval v = st->part()->instrument(inputState().tick())->transpose();
-
-    // If an accidental is set in the input state, use it as a hint for the pitch spelling
-    if (AccidentalType at = m_is.accidentalType(); at == AccidentalType::FLAT || at == AccidentalType::SHARP) {
-        Prefer prefer = at == AccidentalType::SHARP ? Prefer::SHARPS : Prefer::FLATS;
-        if (concertPitch || v.isZero()) {
-            // Note: using Key::C always and ignoring actual key signature. Otherwise, flat mode would still use sharps
-            // sometimes if they're in the key, and vice versa, which seems contrary to the intent of the hint.
-            nval.tpc1 = pitch2tpc(nval.pitch, Key::C, prefer);
-            Interval vFlipped = v;
-            vFlipped.flip();
-            nval.tpc2 = Transpose::transposeTpc(nval.tpc1, vFlipped, true);
-        } else {
-            // Spell the transposed pitch first, then convert to concert pitch
-            int writtenPitch = nval.pitch;
-            if (!allowTransposition) {
-                writtenPitch -= v.chromatic;
-            }
-            nval.tpc2 = pitch2tpc(writtenPitch, Key::C, prefer);
-            nval.tpc1 = Transpose::transposeTpc(nval.tpc2, v, true);
-        }
-    }
-
-    // if transposing, interpret MIDI pitch as representing desired written pitch
-    // set pitch based on corresponding sounding pitch
-    if (!concertPitch && allowTransposition) {
-        nval.pitch += v.chromatic;
-    }
-
-    return nval;
-}
-
-//---------------------------------------------------------
-//  addMidiPitch
-//---------------------------------------------------------
-
-Note* Score::addMidiPitch(int pitch, bool addFlag, bool allowTransposition)
-{
-    NoteVal nval = noteVal(pitch, m_is.staffIdx(), allowTransposition);
-    return addPitch(nval, addFlag);
 }
 
 //---------------------------------------------------------
@@ -3291,28 +3099,6 @@ void Score::cmdDeleteTuplet(Tuplet* tuplet, bool replaceWithRest)
     }
     if (replaceWithRest) {
         setRest(tuplet->tick(), tuplet->track(), tuplet->ticks(), true, tuplet->tuplet());
-    }
-}
-
-//---------------------------------------------------------
-//   nextInputPos
-//---------------------------------------------------------
-
-void Score::nextInputPos(const ChordRest* cr, bool doSelect)
-{
-    ChordRest* ncr = nextChordRest(cr);
-    if ((!ncr) && (m_is.track() % VOICES)) {
-        Segment* s = tick2segment(cr->endTick(), false, SegmentType::ChordRest);
-        ncr = s ? toChordRest(s->element(cr->staffIdx() * VOICES)) : nullptr;
-    }
-    if (ncr) {
-        m_is.setSegment(ncr->segment());
-        if (doSelect) {
-            select(ncr, SelectType::SINGLE, 0);
-        }
-        for (MuseScoreView* v : m_viewer) {
-            v->moveCursor();
-        }
     }
 }
