+++ src/engraving/rw/read460/read460.cpp
@@ -22,6 +22,7 @@
 #include "read460.h"
 
 #include "../editing/mscoreview.h"
+#include "../editing/noteinput.h"
 #include "../editing/transaction/transaction.h"
 #include "../editing/transpose.h"
 #include "../types/types.h"
@@ -350,7 +351,8 @@ bool Read460::preparePasteDurationElement(Score* score, const Fraction& tick, co
     if (Segment* leftSeg = score->tick2leftSegment(tick)) {
         ChordRest* prevCr = leftSeg->nextChordRest(track, /*backwards*/ true, /*stopAtMeasureBoundary*/ true);
         if (prevCr && prevCr->endTick() > tick) {
-            score->truncateChordRest(prevCr, tick, /*fillWithRest*/ false);
+            NoteInput::truncateChordRest(
+                score->transactionManager()->currentOrDummyTransaction(), score, prevCr, tick, /*fillWithRest*/ false);
         }
     }
