+++ src/engraving/rw/read500/read500.cpp
@@ -22,6 +22,7 @@
 #include "read500.h"
 
 #include "../editing/mscoreview.h"
+#include "../editing/noteinput.h"
 #include "../editing/transaction/transaction.h"
 #include "../editing/transpose.h"
 #include "../types/types.h"
@@ -356,7 +357,8 @@ bool Read500::preparePasteDurationElement(Score* score, const Fraction& tick, co
     if (Segment* leftSeg = score->tick2leftSegment(tick)) {
         ChordRest* prevCr = leftSeg->nextChordRest(track, /*backwards*/ true, /*stopAtMeasureBoundary*/ true);
         if (prevCr && prevCr->endTick() > tick) {
-            score->truncateChordRest(prevCr, tick, /*fillWithRest*/ false);
+            NoteInput::truncateChordRest(
+                score->transactionManager()->currentOrDummyTransaction(), score, prevCr, tick, /*fillWithRest*/ false);
         }
     }
