+++ src/engraving/tests/lv_tests.cpp
@@ -35,6 +35,8 @@
 #include "engraving/dom/tremolosinglechord.h"
 
 #include "engraving/editing/edittie.h"
+#include "engraving/editing/noteinput.h"
+#include "engraving/editing/transaction/transaction.h"
 
 #include "utils/scorerw.h"
 
@@ -66,7 +68,7 @@ TEST_F(Engraving_LVTests, LV_Double_Notehead_test)
     score->inputState().setNoteEntryMode(true);
 
     // Add the First Note
-    score->cmdAddPitch(45, true, false);
+    NoteInput::addPitch(score->transactionManager()->currentOrDummyTransaction(), score, 45, true, false);
 
     Measure* m = score->firstMeasure();
     Chord* c = m->findChord(Fraction(0, 1), 0);
@@ -81,7 +83,7 @@ TEST_F(Engraving_LVTests, LV_Double_Notehead_test)
 
     // Add a Second Note
     NoteVal newNoteVal(71);
-    Note* second_note = score->addPitch(newNoteVal, true);
+    Note* second_note = NoteInput::addPitch(score->transactionManager()->currentOrDummyTransaction(), score, newNoteVal, true);
 
     // Check if there is a double notehead
     std::vector<Note*> tn = second_note->tiedNotes();
