+++ src/notation/internal/notationmidiinput.cpp
@@ -29,15 +29,16 @@
 #include "engraving/dom/score.h"
 #include "engraving/dom/note.h"
 #include "engraving/dom/factory.h"
+#include "engraving/editing/noteinput.h"
+#include "engraving/editing/transaction/transaction.h"
 
 #include "playback/playbackcommands.h"
-
-#include "notationtypes.h"
 
 #include "defer.h"
 #include "log.h"
 
 using namespace mu::notation;
+using namespace mu::engraving;
 
 static constexpr int PROCESS_INTERVAL = 20;
 
@@ -131,7 +132,7 @@ void NotationMidiInput::onRealtimeAdvance()
     }
 }
 
-mu::engraving::Score* NotationMidiInput::score() const
+Score* NotationMidiInput::score() const
 {
     IF_ASSERT_FAILED(m_getScore) {
         return nullptr;
@@ -151,7 +152,7 @@ void NotationMidiInput::doProcessEvents()
         return;
     }
 
-    const mu::engraving::Score* sc = score();
+    const Score* sc = score();
     if (!sc || sc->noStaves()) {
         return;
     }
@@ -263,13 +264,13 @@ void NotationMidiInput::addNoteEventsToInputState()
         const muse::midi::Event::Opcode opcode = event.opcode();
 
         if (opcode == muse::midi::Event::Opcode::NoteOn) {
-            NoteVal nval = score()->noteVal(event.note(), staffIdx, useWrittenPitch);
+            NoteVal nval = NoteInput::noteVal(score(), event.note(), staffIdx, useWrittenPitch);
             nval.velocityOverride = event.velocity7();
             notesOn.push_back(nval);
             m_holdingNotesInInputByDuration = true;
         } else if (opcode == muse::midi::Event::Opcode::NoteOff) {
             if (useVelocityAndDuration) {
-                notesOff.push_back(score()->noteVal(event.note(), staffIdx, useWrittenPitch));
+                notesOff.push_back(NoteInput::noteVal(score(), event.note(), staffIdx, useWrittenPitch));
             }
             m_holdingNotesInInputByDuration = false;
         } else if (opcode == muse::midi::Event::Opcode::ControlChange || opcode == muse::midi::Event::Opcode::PitchBend) {
@@ -299,17 +300,17 @@ void NotationMidiInput::addNoteEventsToInputState()
 
 Note* NotationMidiInput::addNoteToScore(const muse::midi::Event& e)
 {
-    mu::engraving::Score* sc = score();
+    Score* sc = score();
 
-    mu::engraving::MidiInputEvent inputEv;
+    MidiInputEvent inputEv;
     inputEv.pitch = e.note();
     inputEv.velocity = e.velocity7();
 
-    sc->activeMidiPitches().remove_if([&inputEv](const mu::engraving::MidiInputEvent& val) {
+    sc->activeMidiPitches().remove_if([&inputEv](const MidiInputEvent& val) {
         return inputEv.pitch == val.pitch;
     });
 
-    const mu::engraving::InputState& is = sc->inputState();
+    const InputState& is = sc->inputState();
     if (!is.noteEntryMode()) {
         return nullptr;
     }
@@ -351,13 +352,14 @@ Note* NotationMidiInput::addNoteToScore(const muse::midi::Event& e)
 
     // holding shift while inputting midi will add the new pitch to the prior existing chord
     if (QGuiApplication::queryKeyboardModifiers() & Qt::ShiftModifier) {
-        mu::engraving::EngravingItem* cr = is.lastSegment()->element(is.track());
+        EngravingItem* cr = is.lastSegment()->element(is.track());
         if (cr && cr->isChord()) {
             inputEv.chord = true;
         }
     }
 
-    mu::engraving::Note* note = sc->addMidiPitch(inputEv.pitch, inputEv.chord, configuration()->midiUseWrittenPitch().val);
+    Note* note = NoteInput::addMidiPitch(sc->transactionManager()->currentOrDummyTransaction(), sc, inputEv.pitch,
+                                                        inputEv.chord, configuration()->midiUseWrittenPitch().val);
 
     sc->activeMidiPitches().push_back(inputEv);
 
@@ -374,21 +376,21 @@ Note* NotationMidiInput::makePreviewNote(const muse::midi::Event& e)
         return nullptr;
     }
 
-    mu::engraving::Score* score = this->score();
-    const mu::engraving::InputState& inputState = score->inputState();
+    Score* score = this->score();
+    const InputState& inputState = score->inputState();
     Segment* seg = inputState.lastSegment() ? inputState.lastSegment() : score->dummy()->segment();
 
     Chord* chord = engraving::Factory::createChord(seg);
     chord->setParent(seg);
 
     const ChordRest* cr = inputState.cr();
-    const mu::engraving::staff_idx_t staffIdx = cr ? engraving::track2staff(cr->track()) : 0;
+    const staff_idx_t staffIdx = cr ? engraving::track2staff(cr->track()) : 0;
 
     Note* note = engraving::Factory::createNote(chord);
     note->setParent(chord);
     note->setStaffIdx(staffIdx);
 
-    engraving::NoteVal nval = score->noteVal(e.note(), staffIdx, configuration()->midiUseWrittenPitch().val);
+    engraving::NoteVal nval = NoteInput::noteVal(score, e.note(), staffIdx, configuration()->midiUseWrittenPitch().val);
     note->setNval(nval);
 
     return note;
@@ -403,7 +405,7 @@ void NotationMidiInput::triggerControllers(const ControllerEventMap& events)
         { muse::midi::SUSTAIN_PEDAL_CONTROLLER, muse::mpe::ControllerChangeEvent::SustainPedalOnOff },
     };
 
-    const mu::engraving::InputState& is = score()->inputState();
+    const InputState& is = score()->inputState();
 
     for (const auto& pair : events) {
         const muse::midi::Event& e = pair.second;
@@ -435,7 +437,7 @@ void NotationMidiInput::releasePlayingNotes(const std::vector<int>& pitches)
     const bool useWrittenPitch = configuration()->midiUseWrittenPitch().val;
 
     for (int pitch : pitches) {
-        const NoteVal nval = score()->noteVal(pitch, staffIdx, useWrittenPitch);
+        const NoteVal nval = NoteInput::noteVal(score(), pitch, staffIdx, useWrittenPitch);
 
         auto it = m_playingNotes.find(nval.pitch);
         if (it == m_playingNotes.end()) {
@@ -503,7 +505,7 @@ void NotationMidiInput::doRealtimeAdvance()
         return;
     }
 
-    const mu::engraving::InputState& is = m_getScore->score()->inputState();
+    const InputState& is = m_getScore->score()->inputState();
     playbackController()->playMetronome(is.tick().ticks());
 
     QTimer::singleShot(100, Qt::PreciseTimer, [this]() {
